+++ src/lib/getCrumb.spec.ts
@@ -1,11 +1,20 @@
 import env from "../env-node";
 import getCrumb, { _getCrumb, getCrumbClear } from "./getCrumb";
-import { MyCookieJar } from "./cookieJar.js";
 import { jest } from "@jest/globals";
 import { consoleSilent, consoleRestore } from "../../tests/console.js";
+import options from "./options.js";
+
+import { ExtendedCookieJar } from "./cookieJar.js";
 
 describe("getCrumb", () => {
-  beforeAll(consoleSilent);
+  const { logger } = options;
+  if (!logger) throw new Error("logger was unset");
+
+  let cookieJar: ExtendedCookieJar;
+  beforeAll(() => {
+    consoleSilent();
+    cookieJar = new ExtendedCookieJar();
+  });
   afterAll(consoleRestore);
 
   describe("_getCrumb", () => {
@@ -13,13 +22,13 @@ describe("getCrumb", () => {
       const fetch = await env.fetchDevel();
 
       const crumb = await _getCrumb(
+        new ExtendedCookieJar(),
         fetch,
-        // @ts-expect-error: fetchDevel still has no types (yet)
         { devel: true },
+        logger,
         "https://finance.yahoo.com/quote/AAPL",
         "getCrumb-quote-AAPL.json",
-        true,
-        new MyCookieJar()
+        true
       );
       expect(crumb).toBe("mloUP8q7ZPH");
     });
@@ -27,11 +36,7 @@ describe("getCrumb", () => {
     it("ditto with shared cookie jar (don't use it for other tests)", async () => {
       const fetch = await env.fetchDevel();
 
-      const crumb = await _getCrumb(
-        fetch,
-        // @ts-expect-error: fetchDevel still has no types (yet)
-        { devel: true }
-      );
+      const crumb = await _getCrumb(cookieJar, fetch, { devel: true }, logger);
       expect(crumb).toBe("mloUP8q7ZPH");
     });
 
@@ -39,9 +44,10 @@ describe("getCrumb", () => {
       const fetch = await env.fetchDevel();
 
       let crumb = await _getCrumb(
+        cookieJar,
         fetch,
-        // @ts-expect-error: fetchDevel still has no types (yet)
         { devel: true },
+        logger,
         "https://finance.yahoo.com/quote/AAPL"
       );
       expect(crumb).toBe("mloUP8q7ZPH");
@@ -49,9 +55,10 @@ describe("getCrumb", () => {
       // TODO, at tests to see how many times fetch was called, etc.
 
       crumb = await _getCrumb(
+        cookieJar,
         fetch,
-        // @ts-expect-error: fetchDevel still has no types (yet)
         { devel: true },
+        logger,
         "https://finance.yahoo.com/quote/AAPL"
       );
       expect(crumb).toBe("mloUP8q7ZPH");
@@ -62,13 +69,13 @@ describe("getCrumb", () => {
 
       await expect(() =>
         _getCrumb(
+          new ExtendedCookieJar(),
           fetch,
-          // @ts-expect-error: fetchDevel still has no types (yet)
           { devel: true },
+          logger,
           "https://finance.yahoo.com/quote/AAPL",
           "getCrumb-quote-AAPL-no-cookies.fake.json",
-          true,
-          new MyCookieJar()
+          true
         )
       ).rejects.toThrowError(/No set-cookie/);
     });
@@ -78,13 +85,13 @@ describe("getCrumb", () => {
 
       await expect(() =>
         _getCrumb(
+          new ExtendedCookieJar(),
           fetch,
-          // @ts-expect-error: fetchDevel still has no types (yet)
           { devel: true },
+          logger,
           "https://finance.yahoo.com/quote/AAPL",
           "getCrumb-quote-AAPL-no-context.fake.json",
-          true,
-          new MyCookieJar()
+          true
         )
       ).rejects.toThrowError(/Could not find window.YAHOO.context/);
     });
@@ -94,13 +101,13 @@ describe("getCrumb", () => {
 
       await expect(() =>
         _getCrumb(
+          new ExtendedCookieJar(),
           fetch,
-          // @ts-expect-error: fetchDevel still has no types (yet)
           { devel: true },
+          logger,
           "https://finance.yahoo.com/quote/AAPL",
           "getCrumb-quote-AAPL-invalid-json.fake.json",
-          true,
-          new MyCookieJar()
+          true
         )
       ).rejects.toThrowError(/Could not parse window.YAHOO.context/);
     });
@@ -110,13 +117,13 @@ describe("getCrumb", () => {
 
       await expect(() =>
         _getCrumb(
+          new ExtendedCookieJar(),
           fetch,
-          // @ts-expect-error: fetchDevel still has no types (yet)
           { devel: true },
+          logger,
           "https://finance.yahoo.com/quote/AAPL",
           "getCrumb-quote-AAPL-no-crumb.fake.json",
-          true,
-          new MyCookieJar()
+          true
         )
       ).rejects.toThrowError(/Could not find crumb/);
     });
@@ -126,13 +133,13 @@ describe("getCrumb", () => {
       const fetch = await env.fetchDevel();
 
       const crumb = await _getCrumb(
+        new ExtendedCookieJar(),
         fetch,
-        // @ts-expect-error: fetchDevel still has no types (yet)
         { devel: true },
+        logger,
         "https://finance.yahoo.com/quote/AAPL",
         "getCrumb-quote-AAPL-pre-consent-VPN-UK.json",
-        true,
-        new MyCookieJar()
+        true
       );
       expect(crumb).toBe("Ky3Po5TGQRZ");
       // consoleSilent();
@@ -141,35 +148,33 @@ describe("getCrumb", () => {
 
   describe("getCrumb", () => {
     it("works", async () => {
-      getCrumbClear();
+      await getCrumbClear(cookieJar);
       const fetch = await env.fetchDevel();
-      const crumb = await getCrumb(
-        fetch,
-        // @ts-expect-error: fetchDevel still has no types (yet)
-        { devel: true }
-      );
+      const crumb = await getCrumb(cookieJar, fetch, { devel: true }, logger);
       expect(crumb).toBe("mloUP8q7ZPH");
     });
 
     it("only calls getCrumb once", async () => {
       const fetch = await env.fetchDevel();
       const _getCrumb = jest.fn(() => "crumb");
-      getCrumbClear();
+      await getCrumbClear(cookieJar);
 
       getCrumb(
+        cookieJar,
         fetch,
-        // @ts-expect-error: fetchDevel still has no types (yet)
         { devel: true },
+        logger,
         "https://finance.yahoo.com/quote/TSLA",
-        _getCrumb
+        _getCrumb as any
       );
 
       getCrumb(
+        cookieJar,
         fetch,
-        // @ts-expect-error: fetchDevel still has no types (yet)
         { devel: true },
+        logger,
         "https://finance.yahoo.com/quote/TSLA",
-        _getCrumb
+        _getCrumb as any
       );
 
       expect(_getCrumb).toHaveBeenCalledTimes(1);