+++ src/lib/setGlobalConfig.ts
@@ -1,11 +1,15 @@
 import type { YahooFinanceOptions } from "./options.js";
 import type { ModuleThis } from "./moduleCommon.js";
 import validateAndCoerceTypes from "./validateAndCoerceTypes.js";
+import { ExtendedCookieJar } from "./cookieJar.js";
 
 export default function setGlobalConfig(
   this: ModuleThis,
-  config: YahooFinanceOptions
+  _config: YahooFinanceOptions
 ): void {
+  // Instances (e.g. cookieJar) don't validate well :)
+  const { cookieJar, logger, ...config } = _config;
+
   validateAndCoerceTypes({
     object: config,
     source: "setGlobalConfig",
@@ -13,15 +17,27 @@ export default function setGlobalConfig(
     options: this._opts.validation,
     schemaKey: "#/definitions/YahooFinanceOptions",
   });
-  mergeObjects(this._opts, config as Obj);
+  mergeObjects(this._opts, config);
+
+  if (cookieJar) {
+    if (!(cookieJar instanceof ExtendedCookieJar))
+      throw new Error("cookieJar must be an instance of ExtendedCookieJar");
+    this._opts.cookieJar = cookieJar;
+  }
+  if (logger) {
+    if (typeof logger.info !== "function")
+      throw new Error("logger.info must be a function");
+    if (typeof logger.warn !== "function")
+      throw new Error("logger.warn must be a function");
+    if (typeof logger.error !== "function")
+      throw new Error("logger.error must be a function");
+    if (typeof logger.debug !== "function")
+      throw new Error("logger.debug must be a function");
+    this._opts.logger = logger;
+  }
 }
 
-type Obj = Record<string, string | ObjRecurse>;
-
-// This is fine, since this is just a hack for recursive types
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-interface ObjRecurse extends Obj {}
-
+type Obj = Record<string, any>;
 function mergeObjects(original: Obj, objToMerge: Obj) {
   const ownKeys: (keyof typeof objToMerge)[] = Reflect.ownKeys(
     objToMerge